+++ churaverse-plugins-client/src/meetingPlugins/meetingCoreUiPlugin/meetingCoreUiPlugin.ts
@@ -7,7 +7,6 @@ import {
 } from 'churaverse-engine-client'
 import { Scene } from 'phaser'
 import { MeetingScreenComponent } from './components/MeetingScreenComponent'
-import { initSidebarToggle } from './components/MeetingSidebarComponent'
 
 export class MeetingCoreUiPlugin extends BasePlugin<IMeetingScene> {
   private scene?: Scene
@@ -23,6 +22,44 @@ export class MeetingCoreUiPlugin extends BasePlugin<IMeetingScene> {
 
   private start(ev: StartEvent): void {
     DomManager.addJsxDom(MeetingScreenComponent())
-    initSidebarToggle()
+    this.initSidebarToggle()
+  }
+
+  private initSidebarToggle(): void {
+    const sidebar = document.getElementById('meeting-sidebar')
+    const participantsSection = document.getElementById('participants-section')
+    const chatSection = document.getElementById('chat-section')
+    const participantsButton = document.getElementById('participants-toggle-button')
+    const chatButton = document.getElementById('chat-toggle-button')
+    const closeButton = document.getElementById('sidebar-close-button')
+    const closeButtonChat = document.getElementById('sidebar-close-button-chat')
+
+    const showSidebar = (tab: 'participants' | 'chat'): void => {
+      if (sidebar == null) return
+      const currentTab = sidebar.getAttribute('data-tab')
+      const isVisible = sidebar.getAttribute('data-visible') === 'true'
+
+      if (isVisible && currentTab === tab) {
+        sidebar.setAttribute('data-visible', 'false')
+      } else {
+        sidebar.setAttribute('data-visible', 'true')
+        sidebar.setAttribute('data-tab', tab)
+
+        if (participantsSection != null && chatSection != null) {
+          participantsSection.setAttribute('data-active', tab === 'participants' ? 'true' : 'false')
+          chatSection.setAttribute('data-active', tab === 'chat' ? 'true' : 'false')
+        }
+      }
+    }
+
+    const closeSidebar = (): void => {
+      if (sidebar == null) return
+      sidebar.setAttribute('data-visible', 'false')
+    }
+
+    participantsButton?.addEventListener('click', () => { showSidebar('participants') })
+    chatButton?.addEventListener('click', () => { showSidebar('chat') })
+    closeButton?.addEventListener('click', () => { closeSidebar() })
+    closeButtonChat?.addEventListener('click', () => { closeSidebar() })
   }
 }
